requirements.txt
discord.py>=2.1.0
python-dotenv>=0.21.1
openai>=0.27.6